+++ NEWS.md
@@ -1,7 +1,8 @@
 # DedooseR 0.0.3
 
 # DedooseR 0.0.2
-* 2025-10-02: Updated the clean_data package to include an exportable codebook!
+* 2025-10-03: Updated the count_codes function to output transcript coverage
+* 2025-10-02: Updated the clean_data function to include an exportable codebook!
 
 # DedooseR 0.0.1